+++ port/nxp/lpc17xx/lpc17_ssp.h
@@ -48,7 +48,7 @@ typedef struct {
 
 void ssp_init( uint8_t id, uint32_t bitrate, uint8_t frame_sz, bool master_mode, bool poll );
 void ssp_ssel_control( uint8_t id, t_ssel_state state );
-void ssp_write_read( uint8_t id, uint8_t *tx_buf, uint8_t tx_len, uint8_t *rx_buf, uint8_t rx_len );
+void ssp_write_read( uint8_t id, uint8_t *tx_buf, uint32_t tx_len, uint8_t *rx_buf, uint32_t rx_len );
 
 #define ssp_chip_init(id)                   Chip_SSP_Init(SSP(id))
 #define ssp_chip_deinit(id)                 Chip_SSP_DeInit(SSP(id))
